+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to Phantom Terminal will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.3.8] - 2026-01-26
+
+### Changed
+- **Browser-style inline suggestions** - Gray text appears as you type
+- Press `RightArrow` or `End` to accept suggestion
+
+---
+
 ## [3.3.7] - 2026-01-26
 
 ### Fixed